Fixing some little mistakes

+++ src/test/java/net/algart/arrays/demo/MainOperationsTest.java
@@ -288,7 +288,7 @@ public class MainOperationsTest implements Cloneable {
             result.mwork1 = null; // for packed combiner
         }
         result.workJA1 = result.a.newJavaArray(result.len);
-        result.workJBits = new long[(int) PackedBitArrays.packedLength(result.len)];
+        result.workJBits = new long[PackedBitArrays.packedLength(result.len)];
         return result;
     }
